Completed Model3 xml comments
