+++ src/DevelopmentApps/apps/DebugApp/DebugApp.cs
@@ -10,7 +10,6 @@ namespace NetDaemon.DevelopmentApps.apps.DebugApp
     /// </summary>
     public class DebugApp : NetDaemonRxApp
     {
-        
         // Use two guids, one when instanced and one when initialized
         // can track errors with instancing
         private Guid _instanceId = Guid.NewGuid();
@@ -27,7 +26,7 @@ namespace NetDaemon.DevelopmentApps.apps.DebugApp
 
         [HomeAssistantServiceCall]
         public void CallMeFromHass(dynamic data)
-             {
+        {
             Log("A call from hass! {data}", data);
         }
 
@@ -41,9 +40,9 @@ namespace NetDaemon.DevelopmentApps.apps.DebugApp
                 Log("State changed as expected");
             }
             catch (System.Exception)
-        {
+            {
                 Log("We had timeout");
-        }
+            }
             await Task.Delay(10);
         }
     }
