+++ src/DevelopmentApps/apps/M3Test/M3App.cs
@@ -2,10 +2,10 @@ using System;
 using System.Reactive.Linq;
 using System.Text.Json;
 using NetDaemon.Common;
+using NetDaemon.DevelopmentApps.Domains.Climate;
+using NetDaemon.DevelopmentApps.Domains.Sensor;
+using NetDaemon.DevelopmentApps.Domains.Zone;
 using NetDaemon.Model3.Common;
-using NetDaemon.Model3.Domains.Climate;
-using NetDaemon.Model3.Domains.Sensor;
-using NetDaemon.Model3.Domains.Zone;
 using NetDaemon.Model3.Entities;
 
 namespace NetDaemon.DevelopmentApps.apps.M3Test
@@ -24,7 +24,7 @@ namespace NetDaemon.DevelopmentApps.apps.M3Test
             // Ha.CallService("notify", "persistent_notification", new { message = "Hello", title = "Yay it works in Model3 via HaContext" }, true);;
 
             _climateEntity = new ClimateEntity(ha, "climate.dummy_thermostat");
-
+            
             _climateEntity.StateAllChanges.Where(e => e.New?.Attributes?.Temperature > 20).Subscribe();
             _climateEntity.StateAllChanges.Subscribe(OnNext);
