+++ src/DevelopmentApps/apps/ZhaApp.cs
@@ -1,8 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Reactive.Linq;
+using System.Text.Json.Serialization;
 using NetDaemon.Common;
 using NetDaemon.Model3.Common;
-using NetDaemon.Model3.EventTypes;
 
 namespace NetDaemon.DevelopmentApps.apps
 {
@@ -41,6 +42,16 @@ namespace NetDaemon.DevelopmentApps.apps
                 }
             }
              */
+        }
+        
+        public record ZhaEventData
+        {
+            [JsonPropertyName("device_ieee")] public string? DeviceIeee { get; set; }
+            [JsonPropertyName("unique_id")] public string? UniqueId { get; set; }
+            [JsonPropertyName("endpoint_id")] public int? EndpointId { get; set; }
+            [JsonPropertyName("endpoint_id")] public int? ClusterId { get; set; }
+            [JsonPropertyName("command")] public string? Command { get; set; }
+            [JsonPropertyName("args")] public IReadOnlyCollection<object>? Args { get; set; }
         }
     }
 }
