+++ src/Model3/NetDeamon.Model3/Common/IEventProvider.cs
@@ -2,10 +2,17 @@
 
 namespace NetDaemon.Model3.Common
 {
+    /// <summary>
+    /// Provides methds for consuming events from Home Assistant
+    /// </summary>
     public interface IEventProvider
     {
-        // TODO: not sure yet if this should be part of IHaCOntext or as a separate interface
-        
+        /// <summary>
+        /// Gets an Observable for a specific eventType and retrieves the EventData deserialized in a specific type
+        /// </summary>
+        /// <param name="eventType">The type of event to retrieve</param>
+        /// <typeparam name="T">The type to deserialize the event data into</typeparam>
+        /// <returns>An IObservable of the filtered Event stream</returns>
         IObservable<T> GetEventDataOfType<T>(string eventType) where T : class;
     }
 }
