+++ src/Model3/NetDeamon.Model3/Internal/ExtensionMethods.cs
