+++ src/Model3/NetDeamon.Model3/NetDaemon.Model3.csproj
@@ -26,4 +26,10 @@
     <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
     <AnalysisMode>AllEnabledByDefault</AnalysisMode>
   </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <DocumentationFile>bin\Debug\NetDaemon.Model3.xml</DocumentationFile>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <DocumentationFile>bin\Release\NetDaemon.Model3.xml</DocumentationFile>
+  </PropertyGroup>
 </Project>
